12: Refactor app
- Put all launch data (pads and launches) in new `missions` folder
- Put favorites in `favorites` folder
- Add firebase-debug log
- Add space X image to be shown at users' sign-up/log in page
