+++ firebase-debug.log
@@ -1,0 +1,44 @@
+[debug] [2020-11-08T17:59:03.687Z] ----------------------------------------------------------------------
+[debug] [2020-11-08T17:59:03.689Z] Command:       C:\DEV\node\node.exe C:\DEV\node\node_modules\firebase-tools\lib\bin\firebase.js login --interactive
+[debug] [2020-11-08T17:59:03.689Z] CLI Version:   8.15.1
+[debug] [2020-11-08T17:59:03.689Z] Platform:      win32
+[debug] [2020-11-08T17:59:03.689Z] Node Version:  v14.13.0
+[debug] [2020-11-08T17:59:03.691Z] Time:          Sun Nov 08 2020 19:59:03 GMT+0200 (Eastern European Standard Time)
+[debug] [2020-11-08T17:59:03.691Z] ----------------------------------------------------------------------
+[debug] [2020-11-08T17:59:03.691Z] 
+[info] i  Firebase optionally collects CLI usage and error reporting information to help improve our products. Data is collected in accordance with Google's privacy policy (https://policies.google.com/privacy) and is not used to identify you.
+ 
+[info] i  To change your data collection preference at any time, run `firebase logout` and log in again. 
+[info] 
+[info] Visit this URL on this device to log in:
+[info] https://accounts.google.com/o/oauth2/auth?client_id=563584335869-fgrhgmd47bqnekij5i8b5pr03ho849e6.apps.googleusercontent.com&scope=email%20openid%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloudplatformprojects.readonly%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Ffirebase%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloud-platform&response_type=code&state=103411119&redirect_uri=http%3A%2F%2Flocalhost%3A9005
+[info] 
+[info] Waiting for authentication...
+[debug] [2020-11-08T17:59:07.367Z] Error: spawn cmd ENOENT
+    at Process.ChildProcess._handle.onexit (internal/child_process.js:268:19)
+    at onErrorNT (internal/child_process.js:464:16)
+    at processTicksAndRejections (internal/process/task_queues.js:80:21)
+[error] 
+[error] Error: An unexpected error has occurred.
+[debug] [2020-11-08T17:59:57.203Z] ----------------------------------------------------------------------
+[debug] [2020-11-08T17:59:57.205Z] Command:       C:\DEV\node\node.exe C:\DEV\node\node_modules\firebase-tools\lib\bin\firebase.js login --interactive
+[debug] [2020-11-08T17:59:57.205Z] CLI Version:   8.15.1
+[debug] [2020-11-08T17:59:57.205Z] Platform:      win32
+[debug] [2020-11-08T17:59:57.206Z] Node Version:  v14.13.0
+[debug] [2020-11-08T17:59:57.207Z] Time:          Sun Nov 08 2020 19:59:57 GMT+0200 (Eastern European Standard Time)
+[debug] [2020-11-08T17:59:57.207Z] ----------------------------------------------------------------------
+[debug] [2020-11-08T17:59:57.207Z] 
+[info] i  Firebase optionally collects CLI usage and error reporting information to help improve our products. Data is collected in accordance with Google's privacy policy (https://policies.google.com/privacy) and is not used to identify you.
+ 
+[info] i  To change your data collection preference at any time, run `firebase logout` and log in again. 
+[info] 
+[info] Visit this URL on this device to log in:
+[info] https://accounts.google.com/o/oauth2/auth?client_id=563584335869-fgrhgmd47bqnekij5i8b5pr03ho849e6.apps.googleusercontent.com&scope=email%20openid%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloudplatformprojects.readonly%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Ffirebase%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloud-platform&response_type=code&state=606254035&redirect_uri=http%3A%2F%2Flocalhost%3A9005
+[info] 
+[info] Waiting for authentication...
+[debug] [2020-11-08T17:59:59.011Z] Error: spawn cmd ENOENT
+    at Process.ChildProcess._handle.onexit (internal/child_process.js:268:19)
+    at onErrorNT (internal/child_process.js:464:16)
+    at processTicksAndRejections (internal/process/task_queues.js:80:21)
+[error] 
+[error] Error: An unexpected error has occurred.
